requirements.txt
# Backend dependencies
flask==3.0.0
flask-cors==4.0.0
aiogram==3.3.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.0

# Production server
gunicorn==21.2.0
python-multipart==0.0.6
